chore(test): Simplify test config files
Merge all test/config/*.js into test/config.js and use the result of credentials.js to configure
tests
