+++ package.json
@@ -15,10 +15,10 @@
     "build": "npm run build:debug && npm run build:release",
     "build:release": "./node_modules/.bin/webpack --release",
     "build:debug": "./node_modules/.bin/webpack",
-    "start": "NODE_ENV=development ./node_modules/.bin/karma start --auto-watch",
+    "start": "./node_modules/.bin/karma start --auto-watch",
     "jwt": "node ./jwt.js",
-    "test": "NODE_ENV=development ./node_modules/.bin/karma start --single-run",
-    "test:coverage": "NODE_ENV=development ./node_modules/.bin/karma start --single-run --coverage",
+    "test": "./node_modules/.bin/karma start --single-run",
+    "test:coverage": "./node_modules/.bin/karma start --single-run --coverage",
     "prepublish-dist": "npm run build",
     "publish-dist": "echo \"machine github.com login $GH_TOKEN\" >> ~/.netrc && git config user.name \"webcomOps\" && git config user.email \"webcom.ops@orange.com\" && git add dist && git commit --no-verify --message \"chore(publish): publish dist [ci skip]\" dist && git push --no-verify origin HEAD:master",
     "semantic-release": "semantic-release pre && npm run publish-dist && npm publish && semantic-release post",
