--- test/config/production.js
@@ -1,5 +1,0 @@
-'use strict';
-
-import configure from './default';
-
-configure();
